requirements.txt
tensorflow-gpu==1.15.2
tensorflow-hub==0.7.0
scikit-learn==0.22.2
matplotlib==3.1.3
bert-tensorflow==1.0.1
sentencepiece==0.1.85
flask==1.1.1
speechrecognition
mutagen
gTTS
pyaudio
